requirements.txt
requests>=2.28.0
qrcode[pil]>=7.4.0
Pillow>=9.3.0
